Feat: 작가 등록 신청 기능 구현

+++ src/main/java/com/onetuks/goguma_bookstore/author_debut/model/AuthorStatics.java
@@ -41,9 +41,9 @@ public class AuthorStatics {
   @Builder
   public AuthorStatics(Author author, Long subscribeCount, Long bookCount, Long restockCount) {
     this.author = author;
-    this.subscribeCount = (Long) Objects.requireNonNullElse(subscribeCount, 0);
-    this.bookCount = (Long) Objects.requireNonNullElse(bookCount, 0);
-    this.restockCount = (Long) Objects.requireNonNullElse(restockCount, 0);
+    this.subscribeCount = Objects.requireNonNullElse(subscribeCount, 0L);
+    this.bookCount = Objects.requireNonNullElse(bookCount, 0L);
+    this.restockCount = Objects.requireNonNullElse(restockCount, 0L);
   }
 
   @Override
